refactor: update manager addresses validation
